requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
attrs==24.2.0
billiard==4.2.1
bitarray==3.0.0
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cytoolz==1.0.0
dlib==19.24.6
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
parsimonious==0.10.0
peewee==3.17.7
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rlp==4.0.1
shellingham==1.5.4
six==1.16.0
toolz==1.0.0
typer==0.12.5
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
web3==7.4.0
websockets==13.1
Werkzeug==3.1.1
yarl==1.17.1
